+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix Temperature Sensor persistence
+
 ## 3.0.0-alpha.91 (2024-10-12)
 
 * (theimo1221) Minor fix in Soda-Handle position update